Basic ordering of task results.
Use 'snooze until' as a proxy for 'show after' and order the results by
the 'show after' time.

If we show tasks after the instant they are created, then we get a
reasonable ordering for active tasks, including ones that return from
being snoozed.

For 'Done' tasks, we probably want to order by the time that they were
marked done, but we can change that later.