Avoid keeping a reference to context in a static
This makes it necessary to remember to (re)schedule a wakeup check after
any change to the task list. It should be possible to handle this
centrally, but to schedule a wakeup requires a context, and I don't
think that TaskData should ideally know about the context and that
world. The Linter agrees that static references to a context are a bad
thing (even though it's the application context).