Cleaner responsibility between notifications and task data
Use DataSetObserver et al for interested parties to be notified when
TaskData changes. Create a new notification manager to be notified when
data changes so it can schedule a new notification run when the next
snoozed tasks wake.

Significantly simplifies the woken task receiver, but there still needs
to be a good way of staring that when the app is installed, updated or
the device restarts.

Means that UI classes don't have to remember to schedule a snooze check
every time they make a TaskData call to update the DB.