Hacky way to display task detail result snackbar after notification.
Normally when TaskDetailActivity finishes, it returns to MainActivity
and, since it was started with startActivityForResult, we get to handle
the result.

Notifications launch TaskDetailActivity with a synthetic back stack, and
there's no 'startActivityForResult' in that mix, so MainActivity doesn't
get to see the results.

So this commit adds a static variable on TaskDetailActivity that
is set to the last result value, that can be picked up by MainActivity
(or others) no matter how it is returned to.