Tidy up CustomSpinnerAdapter API & fix onTimeSet problems
CustomSpinnerAdapter now deals in items (as per underlying Adapter) and
values of a certain type (i.e. the thing picked as a custom value).
Functions that convert from Items to Values can be provided to the
adapter, making it easier to find the position of a value, rather than
of an item.

Tidied up other TODOs, lint errors and cleaned up naming a little.